Moved ts-node to root and removed it from tokens package
