+++ packages/tokens/package.json
@@ -17,7 +17,6 @@
   "devDependencies": {
     "@tokens-studio/sd-transforms": "^0.8.5",
     "change-case": "^4.1.2",
-    "style-dictionary": "^3.7.2",
-    "ts-node": "^10.9.1"
+    "style-dictionary": "^3.7.2"
   }
 }
