+++ packages/tokens/tsconfig.json
@@ -1,10 +1,3 @@
 {
-  "extends": "../../tsconfig.json",
-  "ts-node": {
-    "transpileOnly": true,
-    "files": true,
-    "compilerOptions": {
-      "module": "CommonJS"
-    }
-  }
+  "extends": "../../tsconfig.json"
 }
