+++ scripts/generate-component/fileTemplates.ts
