+++ scripts/generate-component/storefront.ts
@@ -1,27 +1,30 @@
 const componentName = process.argv[2];
 const componentPath = './';
 import { mkdir, writeFile } from 'fs/promises';
-import { mainContent, exportContent, cssContent } from './fileTemplates.mjs';
+
+import { mainContent, exportContent, cssContent } from './fileTemplates';
 
 (async () => {
   // Component folder
-  mkdir(`${componentPath}/${componentName}`).catch(console.error);
+  await mkdir(`${componentPath}/${componentName}`).catch(console.error);
 
   // Component file
-  writeFile(
+  await writeFile(
     `${componentPath}/${componentName}/${componentName}.tsx`,
     mainContent(componentName),
   );
 
   // Css file
-  writeFile(
+  await writeFile(
     `${componentPath}/${componentName}/${componentName}.module.css`,
     cssContent(componentName),
   );
 
   // Export file
-  writeFile(
+  await writeFile(
     `${componentPath}/${componentName}/index.ts`,
     exportContent(componentName),
   );
-})();
+})().catch((e) => {
+  console.log(e);
+});
