+++ tsconfig.json
@@ -25,5 +25,12 @@
     "resolveJsonModule": true,
     "target": "es2016",
     "plugins": [{ "name": "typescript-plugin-css-modules" }]
+  },
+  "ts-node": {
+    "transpileOnly": true,
+    "files": true,
+    "compilerOptions": {
+      "module": "CommonJS"
+    }
   }
 }
